+++ src/flask/globals.py
@@ -14,25 +14,6 @@ if t.TYPE_CHECKING:  # pragma: no cover
     from .wrappers import Request
 
 
-class _FakeStack:
-    def __init__(self, name: str, cv: ContextVar[t.Any]) -> None:
-        self.name = name
-        self.cv = cv
-
-    @property
-    def top(self) -> t.Any | None:
-        import warnings
-
-        warnings.warn(
-            f"'_{self.name}_ctx_stack' is deprecated and will be removed in Flask 2.4."
-            f" Use 'g' to store data, or '{self.name}_ctx' to access the current"
-            " context.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return self.cv.get(None)
-
-
 _no_app_msg = """\
 Working outside of application context.
 
@@ -41,7 +22,6 @@ the current application. To solve this, set up an application context
 with app.app_context(). See the documentation for more information.\
 """
 _cv_app: ContextVar[AppContext] = ContextVar("flask.app_ctx")
-__app_ctx_stack = _FakeStack("app", _cv_app)
 app_ctx: AppContext = LocalProxy(  # type: ignore[assignment]
     _cv_app, unbound_message=_no_app_msg
 )
@@ -60,7 +40,6 @@ an active HTTP request. Consult the documentation on testing for
 information about how to avoid this problem.\
 """
 _cv_request: ContextVar[RequestContext] = ContextVar("flask.request_ctx")
-__request_ctx_stack = _FakeStack("request", _cv_request)
 request_ctx: RequestContext = LocalProxy(  # type: ignore[assignment]
     _cv_request, unbound_message=_no_req_msg
 )
@@ -70,27 +49,3 @@ request: Request = LocalProxy(  # type: ignore[assignment]
 session: SessionMixin = LocalProxy(  # type: ignore[assignment]
     _cv_request, "session", unbound_message=_no_req_msg
 )
-
-
-def __getattr__(name: str) -> t.Any:
-    if name == "_app_ctx_stack":
-        import warnings
-
-        warnings.warn(
-            "'_app_ctx_stack' is deprecated and will be removed in Flask 2.4.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return __app_ctx_stack
-
-    if name == "_request_ctx_stack":
-        import warnings
-
-        warnings.warn(
-            "'_request_ctx_stack' is deprecated and will be removed in Flask 2.4.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return __request_ctx_stack
-
-    raise AttributeError(name)