+++ src/flask/signals.py
@@ -1,8 +1,5 @@
 from __future__ import annotations
 
-import typing as t
-import warnings
-
 from blinker import Namespace
 
 # This namespace is only for signals provided by Flask itself.
@@ -18,16 +15,3 @@ appcontext_tearing_down = _signals.signal("appcontext-tearing-down")
 appcontext_pushed = _signals.signal("appcontext-pushed")
 appcontext_popped = _signals.signal("appcontext-popped")
 message_flashed = _signals.signal("message-flashed")
-
-
-def __getattr__(name: str) -> t.Any:
-    if name == "signals_available":
-        warnings.warn(
-            "The 'signals_available' attribute is deprecated and will be removed in"
-            " Flask 2.4. Signals are always available.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return True
-
-    raise AttributeError(name)